send-to-notion-1/requirements.txt
functions-framework==3.*
flask==2.3.2
google_cloud_storage==2.9.0